// ui/src/generated/MidiMapTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AutomationTarget } from "./AutomationTarget";
import type { TrackId } from "./TrackId";
import type { TransportAction } from "./TransportAction";

export type MidiMapTarget = { "type": "Param", target: AutomationTarget, } | { "type": "TrackMute", track: TrackId, } | { "type": "TrackSolo", track: TrackId, } | { "type": "TrackArm", track: TrackId, } | { "type": "Transport", action: TransportAction, };